melihat skalaNilai resp-service
